pyproject.toml
[tool.poetry]
name = "mss"
version = "0.1.0"
description = ""
authors = ["roma <roma.marhatok@yandex.by>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.5"
load-dotenv = "^0.1.0"
Pillow = "^9.3.0"
Faker = "^15.3.2"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
requests = "^2.28.1"
exrex = "^0.10.5"
Django = "^4.1.5"
regex = "^2023.3.22"
unidecode = "^1.3.6"
faker-file = {extras = ["docx"], version = "^0.12.4"}

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^22.8.0"
pre-commit = "^2.20.0"
flake8-bugbear = "^22.8.23"
pep8-naming = "^0.13.2"
flake8-variables-names = "^0.0.5"
flake8-functions-names = "^0.3.0"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
